moved mvc- and rest-related controllers and configs to dedicated packages
