+++ src/main/java/com/patiun/libraryspring/rest/configuration/RestSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.patiun.libraryspring.configuration;
+package com.patiun.libraryspring.rest.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
