+++ src/main/java/com/patiun/libraryspring/rest/controller/BookRestController.java
@@ -1,7 +1,10 @@
-package com.patiun.libraryspring.book;
+package com.patiun.libraryspring.rest.controller;
 
-import com.patiun.libraryspring.configuration.RestSecurityConfig;
+import com.patiun.libraryspring.book.Book;
+import com.patiun.libraryspring.book.BookEditDto;
+import com.patiun.libraryspring.book.BookService;
 import com.patiun.libraryspring.exception.ServiceException;
+import com.patiun.libraryspring.rest.configuration.RestSecurityConfig;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
